+++ core/XenodeRuntime/src/MainRuntimeLayer.h
@@ -11,7 +11,7 @@ public:
 
 	void OnAttach() override;
 	void OnDetach() override;
-	void OnUpdate(double timestep) override;
+	void OnUpdate(float timestep) override;
 	void OnFixedUpdate() override;
 	void OnRender() override;
